pyproject.toml
[tool.poetry]
name = "bot_carga_solar"
version = "0.1.0"
description = ""
authors = ["LeeboyOver9000 <pedro.ivo.freire.aragao@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
selenium = "^4.1.0"
webdriver-manager = "^3.5.2"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
pytest = "^7.1"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
